// src/model/incident.ts
/**
 * API BHB Customer Backend
 * API BHB Customer Backend
 *
 * OpenAPI spec version: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Incident information
 */
export interface Incident { 
    /**
     * Status creation date
     */
    createdAt?: string;
    /**
     * Status created by
     */
    createdBy?: string;
    /**
     * Incident comments
     */
    comments?: string;
    /**
     * Incident fault
     */
    incidentType?: Incident.IncidentTypeEnum;
    /**
     * Coupon has been generated
     */
    couponGenerated?: boolean;
    /**
     * Incident has been solved
     */
    solved?: boolean;
    /**
     * Order id
     */
    orderId?: string;
}
export namespace Incident {
    export type IncidentTypeEnum = 'Kitchen' | 'Delivery' | 'Client';
    export const IncidentTypeEnum = {
        Kitchen: 'Kitchen' as IncidentTypeEnum,
        Delivery: 'Delivery' as IncidentTypeEnum,
        Client: 'Client' as IncidentTypeEnum
    };
}